requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# codeTrial/urls.py
# codeTrial/wsgi.py
# coderunner/admin.py
# coderunner/apps.py
# coderunner/forms.py
# coderunner/models.py
# coderunner/signals.py
# coderunner/tokens.py
# coderunner/urls.py
# coderunner/views.py
# manage.py
Django == 2.1.2

djangorestframework == 3.8.2

# coderunner/views.py
pylint == 2.1.1

gunicorn >= 19.9.0
dj-database-url == 0.5.0
psycopg2 == 2.7.5
python-decouple >= 3.1
whitenoise >= 4.1
Unipath >= 1.1
